Move explorer to official dir
