+++ .github/workflows/test-integration.yml
@@ -23,7 +23,7 @@ jobs:
       - uses: actions/checkout@v3.4.0
       - name: Finding files and store to output
         id: set-matrix
-        run: echo "matrix=$({ cd integration && find . -type d ! -name testdata -maxdepth 1 -print; } | tail -n +2 | cut -c 3- | jq -R . | jq -cs .)" >> $GITHUB_OUTPUT
+        run: echo "matrix=$({ find . -type d -name integration -print; } | cut -c 3- | jq -R . | jq -cs .)" >> $GITHUB_OUTPUT
 
   integration:
     name: test ${{ matrix.test-path }}
@@ -41,10 +41,6 @@ jobs:
         with:
           PATTERNS: |
             ${{ matrix.test-path }}/**
-            !${{ matrix.test-path }}/README.md
-            integration/go.mod
-            integration/go.sum
-            integration/${{ matrix.test-path }}/**
 
       - uses: actions/setup-go@v4
         if: env.GIT_DIFF
@@ -56,7 +52,7 @@ jobs:
         env:
           GOTOOLCHAIN: local+path
           GOSUMDB: off
-        run: go test -v -timeout 120m ./integration/${{ matrix.test-path }}
+        run: go test -v -timeout 120m ./${{ matrix.test-path }}
 
   status:
     runs-on: ubuntu-latest
