+++ go.work.example
@@ -1,8 +1,7 @@
 go 1.21.1
 
 use (
-	./explorer
+	./official/explorer
 	./hermes
-	./integration
 	./official/marketplace
 )
