--- integration/tools.go
@@ -1,10 +1,0 @@
-//go:build tools
-
-package integration
-
-import (
-	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
-	_ "golang.org/x/tools/cmd/goimports"
-	_ "golang.org/x/vuln/cmd/govulncheck"
-	_ "mvdan.cc/gofumpt"
-)
