+++ official/explorer/integration/gex_test.go
@@ -1,4 +1,4 @@
-package explorer_test
+package integration_test
 
 import (
 	"os"
